requirements.txt
crowdstrike-falconpy>=1.4.0
okta>=2.0.0
git+https://github.com/TheJumpCloud/jcapi-python.git#subdirectory=jcapiv2
pan-os-python
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn
apscheduler>=3.10.0
rapidfuzz
structlog
python-dotenv
jinja2
pytest
pytest-asyncio
httpx
openpyxl
openai
